Lumen-inventaris

+++ app/Http/Controllers/InboundStuffController.php
@@ -11,6 +11,11 @@ use App\Helpers\ApiFormatter;
 
 class InboundStuffController extends Controller
 {
+    public function __construct()
+{
+    $this->middleware('auth:api');
+}
+
     public function index(Request $request)
     {
         try{
@@ -80,38 +85,67 @@ class InboundStuffController extends Controller
         }
     }
 
-    public function destroy($id)
+//     public function destroy($id)
+//     {
+//         try {
+//             $inboundData = InboundStuff::where('id', $id)->first();
+//             //simpan data dari inbound yang diperulukan
+//             $stuffId = $inboundData['stuff_id'];
+//             $totalInbound = $inboundData['total'];
+//             $inboundData->delete();
+            
+
+//             //kurangi total_available sblmny dgn total dr inbound yg akan di hps
+//             $dataStock = StuffStock::where('stuff_id', $inboundData['stuff_id'])->frist();
+//             $total_available = (int)$inboundData['total_available'] - (int)$inboundData['total'];
+            
+//             $minusTotalStock = StuffStock::where('stuff_id', $inboundData['stuff_id'])->update(['total_available' => $total_available]);
+
+//             if ($minusTotalStock) {
+//                 $updatedStuffWithInboundAndStock = Stuff::where('id', $inboundData['stuff_id'])->with('inboundStuffs', 'stuffStock')->first();
+
+//                 // dalete inbound terakhir agar data stuff_id di inbound bisa digunakan untuk mengambil data terbaru
+               
+//                 return ApiFormatter::sendResponse(200, 'success', $updatedStuffWithInboundAndStock);
+//             }
+//         } catch (\Exception $err) {
+//             return ApiFormatter::sendResponse(400, 'bad request', $err->getMessage());
+//         }
+// }
+
+public function destroy($id)
     {
         try {
-            $inboundData = InboundStuff::where('id', $id)->first();
-            //simpan data dari inbound yang diperulukan
+            $inboundData = InboundStuff::where('id', $id)->with('stuff.stuffStock')->first();
+            //simpan data dr inbound yg diperluin setelah di delete
             $stuffId = $inboundData['stuff_id'];
             $totalInbound = $inboundData['total'];
-            $inboundData->delete();
-            
 
-            //kurangi total_available sblmny dgn total dr inbound yg akan di hps
-            $dataStock = StuffStock::where('stuff_id', $inboundData['stuff_id'])->frist();
-            $total_available = (int)$inboundData['total_available'] - (int)$inboundData['total'];
-            
-            $minusTotalStock = StuffStock::where('stuff_id', $inboundData['stuff_id'])->update(['total_available' => $total_available]);
+            //kurangi total avail sebelumnya dengan total dr inbound di hps
+            $dataStock = StuffStock::where('stuff_id', $inboundData['stuff_id'])->first();
+            $total_available = (int)$dataStock['total_available'] - (int)$totalInbound;
 
-            if ($minusTotalStock) {
-                $updatedStuffWithInboundAndStock = Stuff::where('id', $inboundData['stuff_id'])->with('inboundStuffs', 'stuffStock')->first();
-
-                // dalete inbound terakhir agar data stuff_id di inbound bisa digunakan untuk mengambil data terbaru
-               
-                return ApiFormatter::sendResponse(200, 'success', $updatedStuffWithInboundAndStock);
+            if ($total_available < $totalInbound) {
+                return ApiFormatter::sendResponse(400, 'bad request', 'Jumlah total inbound yang akan dihapus lebih besar dari total available stuff ini !');
+            } else {
+                $inboundData->delete();
+                $minusTotalStock = $dataStock->update(['total_available' => $total_available]);
+                if ($minusTotalStock){
+                    $updatedStuffWithInboundAndStock = Stuff::where('id', $inboundData['staff_id'])->with('inboundStuffs','stuffStock')->first();
+                    
+                    $inboundData->delete();
+                    return ApiFormatter::sendResponse(200,'Success',$updatedStuffWithInboundAndStock);
+                }
             }
         } catch (\Exception $err) {
             return ApiFormatter::sendResponse(400, 'bad request', $err->getMessage());
         }
-}
+    }
 
 public function trash() {
     try {
         // onlyTrashed() : memanggil data sampah yang sudah di hapus/ deleted_at nya terisi
-        $data = User::onlyTrashed()->get();
+        $data = InboundStuff::onlyTrashed()->get();
 
             return ApiFormatter::sendResponse(200, 'success', $data);
     } catch (\Exception $err) {
@@ -119,40 +153,53 @@ public function trash() {
     }
 }
 
-public function restore($id) {
+public function restore($id)
+{
     try {
-        // restore : mengembalikan data spesifik yang dihapus/menghapus deleted_at nya
-        $checkRestore = User::onlyTrashed()->where('id', $id)->restore();
+        $restoredRows = InboundStuff::onlyTrashed()->where('id', $id)->restore();
 
-        if ($checkRestore) {
-            $restoredUser = User::where('id', $id)->first();
+        if ($restoredRows > 0) {
+            $inboundData = InboundStuff::find($id);
+            $stuffId = $inboundData['stuff_id'];
+            $totalInbound = $inboundData['total'];
+            $inboundData->delete();
 
-            $inboundStuffs = InboundStuff::onlyTrashed()->where('user_id', $restoredUser->id)->get();
+            // Tambahkan total_available untuk mengembalikan total seperti sebelumnya
+            $dataStock = StuffStock::where('stuff_id', $stuffId)->first();
+            $total_available = (int) $dataStock['total_available'] += (int) $totalInbound;
+            $plusTotalStock = $dataStock->update(['total_available' => $total_available]);
 
-            foreach ($inboundStuffs as $inboundStuff) {
-                $stuffStock = StuffStock::where('stuff_id', $inboundStuff->stuff_id)->first();
-                $stuffStock->total_available += $inboundStuff->quantity;
-                $stuffStock->save();
-            }
+            if ($plusTotalStock) {
+                $updatedStuff = Stuff::where('id', $stuffId)
+                ->with('inboundStuffs', 'stuffStock')
+                ->first();
 
-            $data = User::where('id', $id)->first();
-            return ApiFormatter::sendResponse(200, 'success', $data);
+            return ApiFormatter::sendResponse(200, "success", $updatedStuff);
         }
-    } catch (\Exception $err) {
-        return ApiFormatter::sendResponse(400, 'bad request', $err->getMessage());
     }
+} catch (\Exception $err) {
+    return ApiFormatter::sendResponse(400, "bad request", $err->getMessage());
 }
+}
+
 
 public function permanentDelete($id) {
     try {
-        //->forceDelete() : menghapus permanent (hilangn juga data di db nya)
-        $checkPermanentDelete = User::onlyTrashed()->where('id', $id)->forceDelete();
+        // Ambil data inbound_stuffs yang akan dihapus
+        $inboundStuff = InboundStuff::onlyTrashed()->where('id', $id)->first();
+
+        // Hapus file terkait (misalnya, file upload)
+        Storage::delete($inboundStuff->file_path);
+
+        // Hapus data secara permanen dari basis data
+        $checkPermanentDelete = $inboundStuff->forceDelete();
 
         if ($checkPermanentDelete) {
-            return ApiFormatter::sendResponse(200, 'success', "Berhasil menghapus permanent data User!");
+            return ApiFormatter::sendResponse(200, 'success', "Berhasil menghapus permanen data Inbound Stuff dan file terkait!");
         }
     } catch (\Exception $err) {
         return ApiFormatter::sendResponse(400, 'bad request', $err->getMessage());
     }
 }
+
 }
